+++ tests/testthat/test-ps_calc_diversity.R
@@ -1,5 +1,4 @@
 
-
 test_that("ps_calc_diversity supported plot doesn't change", {
   p <- corncob::ibd_phylo %>%
     ps_filter(abx == "abx") %>%